DigiMixer: Fix bugs, and create a working client
This client receives fader changes (and possibly others).

+++ DigiMixer/DigiMixer.DmSeries.Core/DmUInt16Segment.cs
@@ -5,7 +5,7 @@ namespace DigiMixer.DmSeries.Core;
 
 public sealed class DmUInt16Segment : DmSegment
 {
-    public override DmSegmentFormat Format => DmSegmentFormat.Int32;
+    public override DmSegmentFormat Format => DmSegmentFormat.UInt16;
 
     public override int Length => 5 + Values.Count * 2;
